LayoutTests/http/tests/security/xssAuditor/anchor-url-dom-write-location-inline-event-null-char-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

CONSOLE MESSAGE: Fragment navigation not allowed with cross-origin frames.
CONSOLE MESSAGE: Fragment navigation not allowed with cross-origin frames.
